💄 Add social networks links
 #26
