+++ src/components/Header/Header.css
@@ -1,0 +1,18 @@
+.header {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  background: rebeccapurple;
+  margin-bottom: 1.25rem;
+  padding: 1.5rem 1rem;
+}
+
+.header > h1 {
+  margin: 0;
+  margin-right: auto;
+}
+
+.header > h1 > a {
+  color: white;
+  text-decoration: none;
+}
