+++ src/components/Header/SocialLinks.js
@@ -1,0 +1,20 @@
+import React from "react"
+
+import logoTwitter from "../../images/icons/icon-twitter.svg"
+import logoGithub from "../../images/icons/icon-github.svg"
+import "./SocialLinks.css"
+
+const SocialLinks = () => {
+  return (
+    <div className="social-links-wrapper">
+      <a href="https://twitter.com/ZenikaIT">
+        <img src={logoTwitter} className="social-link-icon" alt="twitter logo"/>
+      </a>
+      <a href="https://github.com/zenika-open-source">
+        <img src={logoGithub} className="social-link-icon" alt="github logo"/>
+      </a>
+    </div>
+  )
+}
+
+export default SocialLinks
