+++ src/components/layout.js
@@ -9,7 +9,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { StaticQuery, graphql } from "gatsby";
 
-import Header from "./header";
+import Header from "./Header/Header";
 import "./layout.css";
 
 const Layout = ({ children }) => (
@@ -36,7 +36,7 @@ const Layout = ({ children }) => (
         >
           <main>{children}</main>
           <footer>
-            <div class="partners">
+            <div className="partners">
               <div>
                 <a href="https://www.netlify.com">
                   <img
@@ -46,7 +46,7 @@ const Layout = ({ children }) => (
                 </a>
               </div>
             </div>
-            <div class="copyright">
+            <div className="copyright">
               © {new Date().getFullYear()}, Built with{" "}
               <a href="https://www.gatsbyjs.org">Gatsby</a>
             </div>
